Modify events and add GameEventListener
Move event so classes and listener to their own files.
add GameEventListener.
