+++ Assets/_Project/ScriptableObjects/Events/EnemySpawnedEventSO.asset
@@ -9,8 +9,8 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: 444dec76264f40f4a8c47e2fbe6b0aa4, type: 3}
   m_Name: EnemySpawnedEventSO
-  m_EditorClassIdentifier: PetalsOfHope.Runtime:PetalsOfHope.Core.Events:GameObjectEventSO
+  m_EditorClassIdentifier: 
   _developerDescription: 'Raised when a new enemy is spawned. Payload: the spawned
     enemy GameObject.'
