+++ Assets/_Project/ScriptableObjects/Events/ShowNotificationEventSO.asset
@@ -9,8 +9,8 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: 2a73011672604aa7af4a99a76b7090eb, type: 3}
   m_Name: ShowNotificationEventSO
-  m_EditorClassIdentifier: PetalsOfHope.Runtime:PetalsOfHope.Core.Events:StringEventSO
+  m_EditorClassIdentifier: 
   _developerDescription: 'Raised to show a notification to the player. Payload: message
     to display.'
