+++ Assets/_Project/ScriptableObjects/Events/ToggleSettingEventSO.asset
@@ -9,8 +9,8 @@ MonoBehaviour:
   m_GameObject: {fileID: 0}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 0}
+  m_Script: {fileID: 11500000, guid: 06da8ee14c7b4d0b9ded6ac8bdaec7c7, type: 3}
   m_Name: ToggleSettingEventSO
-  m_EditorClassIdentifier: PetalsOfHope.Runtime:PetalsOfHope.Core.Events:BoolEventSO
+  m_EditorClassIdentifier: 
   _developerDescription: 'Raised when a game setting is toggled. Payload: new state
     of the setting (true/false).'
