+++ Assets/_Project/Scripts/Core/Events/BoolEventSO.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+namespace PetalsOfHope.Core.Events
+{
+    [CreateAssetMenu(menuName = "Petals of Hope/Events/Typed Events/Bool Event", fileName = "NewBoolEventSO")]
+    public class BoolEventSO : TypedEventSO<bool> {}
+}
