+++ Assets/_Project/Scripts/Core/Events/BoolEventSO.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 06da8ee14c7b4d0b9ded6ac8bdaec7c7
+timeCreated: 1748949296
