+++ Assets/_Project/Scripts/Core/Events/FloatEventSO.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+namespace PetalsOfHope.Core.Events
+{
+    [CreateAssetMenu(menuName = "Petals of Hope/Events/Typed Events/Float Event", fileName = "NewFloatEventSO")]
+    public class FloatEventSO : TypedEventSO<float> {}
+}
