+++ Assets/_Project/Scripts/Core/Events/FloatEventSO.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: ab412caa604544a7a5ae11c75025d118
+timeCreated: 1748949265
