+++ Assets/_Project/Scripts/Core/Events/GameObjectEventSO.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+namespace PetalsOfHope.Core.Events
+{
+    [CreateAssetMenu(menuName = "Petals of Hope/Events/Typed Events/GameObject Event", fileName = "NewGameObjectEventSO")]
+    public class GameObjectEventSO : TypedEventSO<GameObject> {}
+}
