+++ Assets/_Project/Scripts/Core/Events/GameObjectEventSO.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 444dec76264f40f4a8c47e2fbe6b0aa4
+timeCreated: 1748949328
