+++ Assets/_Project/Scripts/Core/Events/IntEventSO.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+namespace PetalsOfHope.Core.Events
+{
+    [CreateAssetMenu(menuName = "Petals of Hope/Events/Typed Events/Int Event", fileName = "NewIntEventSO")]
+    public class IntEventSO : TypedEventSO<int> {}
+}
