+++ Assets/_Project/Scripts/Core/Events/IntEventSO.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 23efa00040d94ae3b7a24bbc2b2408b5
+timeCreated: 1748948621
