+++ Assets/_Project/Scripts/Core/Events/Listeners/BaseTypedEventListener.cs
@@ -1,0 +1,6 @@
+﻿using UnityEngine;
+
+namespace PetalsOfHope.Core.Events.Listeners
+{
+    public abstract class BaseTypedEventListener : MonoBehaviour {}
+}
