+++ Assets/_Project/Scripts/Core/Events/Listeners/BaseTypedEventListener.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 30ffbca28a174e82b315f9bdffa36c8b
+timeCreated: 1748954606
