+++ Assets/_Project/Scripts/Core/Events/Listeners/BoolEventListener.cs
@@ -1,0 +1,9 @@
+﻿using UnityEngine;
+using UnityEngine.Events;
+
+namespace PetalsOfHope.Core.Events.Listeners
+{
+    [AddComponentMenu("Petals of Hope/Event Listeners/Bool Event Listener")]
+    public class BoolEventListener : TypedEventListener<BoolEventSO, UnityBoolEvent, bool> {}
+    [System.Serializable] public class UnityBoolEvent : UnityEvent<bool> {}
+}
