+++ Assets/_Project/Scripts/Core/Events/Listeners/GameEventListener.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 2d8338d9c759bc94292813011fd243a8
