+++ Assets/_Project/Scripts/Core/Events/Listeners/StringEventListener.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 4f84d7d30b654be98adaa0de298b2161
+timeCreated: 1748953957
