+++ Assets/_Project/Scripts/Core/Events/Listeners/Vector2EventListener.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 8ddf67522e5145b0b251e1df39439ba0
+timeCreated: 1748954004
