+++ Assets/_Project/Scripts/Core/Events/StringEventSO.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+namespace PetalsOfHope.Core.Events
+{
+    [CreateAssetMenu(menuName = "Petals of Hope/Events/Typed Events/String Event", fileName = "NewStringEventSO")]
+    public class StringEventSO : TypedEventSO<string> {}
+}
