+++ Assets/_Project/Scripts/Core/Events/StringEventSO.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 2a73011672604aa7af4a99a76b7090eb
+timeCreated: 1748949286
