+++ Assets/_Project/Scripts/Core/Events/Vector2EventSO.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+namespace PetalsOfHope.Core.Events
+{
+    [CreateAssetMenu(menuName = "Petals of Hope/Events/Typed Events/Vector2 Event", fileName = "NewVector2EventSO")]
+    public class Vector2EventSO : TypedEventSO<Vector2> {}
+}
