+++ Assets/_Project/Scripts/Core/Events/Vector2EventSO.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 067ceacfca4c4729854774b50d7bd614
+timeCreated: 1748949303
