+++ Assets/_Project/Scripts/Core/Events/Vector3EventSO.cs
@@ -1,0 +1,7 @@
+﻿using UnityEngine;
+
+namespace PetalsOfHope.Core.Events
+{
+    [CreateAssetMenu(menuName = "Petals of Hope/Events/Typed Events/Vector3 Event", fileName = "NewVector3EventSO")]
+    public class Vector3EventSO : TypedEventSO<Vector3> {}
+}
