+++ Assets/_Project/Scripts/Core/Events/Vector3EventSO.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 9d8aec4bff6d48bbac30e7d115510a4d
+timeCreated: 1748949321
